gcp_fs_cs_bq/cloud_functions/stg_to_dwh/requirements.txt
functions-framework==3.*
google-cloud-firestore
requests
pandas
numpy
google-cloud-bigquery
google-cloud-storage